Fix reviewer comments
