+++ src/main/kotlin/ArchiveMenu.kt
@@ -1,46 +1,16 @@
-import java.util.*
-
 class ArchiveMenu(val name: String) : MenuItem() {
+    override val menuTitle: String = "Archive $name"
+    override val menuItemName: String = "note"
+
     override fun toString(): String {
         return name
     }
 
-    override fun showMenu() {
-        println("==================")
-        println("Archive \"$name\" menu:")
-        println("==================")
-        println("1. Create note")
-        println("2. Remove note")
-        println("3. Show available notes")
-        println("4. Select note")
-        println("5. Exit")
-    }
-
-    override fun handleMenu() {
-        var validInput = false
-        while (!validInput){
-            val input = Scanner(System.`in`).nextLine().toIntOrNull()
-                ?.let {
-                        value ->
-                    when(value) {
-                        1 -> add()
-                        2 -> remove()
-                        3 -> show()
-                        4 -> select()
-                        5 -> return
-                        else -> println("Invalid input, try again")
-                    }
-                    showMenu()
-                }
-                ?: println("Invalid input, try again")
-        }
-    }
-
     override fun add() {
         println("Enter note title")
-        super.getText()
+        getText()
             ?. let { title ->
-                super.addItem(Note(title))
+                addItem(NoteMenu(title))
                 println("Added note \"$title\"")
             }
             ?: println("Invalid note title")
@@ -48,9 +18,9 @@ class ArchiveMenu(val name: String) : MenuItem() {
 
     override fun remove() {
         println("Enter note index")
-        super.getIndex()
+        getIndex()
             ?. let { index ->
-                super.removeItem(index)
+                removeItem(index)
                 println("Note removed")
             }
             ?: println("Invalid note index")
@@ -58,13 +28,13 @@ class ArchiveMenu(val name: String) : MenuItem() {
 
     override fun select() {
         println("Enter note index")
-        super.getIndex()
-            ?. let { index -> super.selectItem(index) }
+        getIndex()
+            ?. let { index -> selectItem(index) }
             ?: println("Invalid note index")
     }
 
     override fun show(){
         println("Notes:")
-        super.showItems()
+        showItems()
     }
 }
