+++ src/main/kotlin/Main.kt
@@ -1,4 +1,3 @@
 fun main() {
-    var menu: ArchiveListMenu = ArchiveListMenu()
-    menu.touch()
+    ArchiveListMenu().touch()
 }
